# openapi.yaml
openapi: 3.0.3
info:
  title: RAG AI 知识库系统 API
  description: |
    基于 RAG 技术的个人知识库 AI 问答系统
    
    ## 认证说明
    除登录/注册接口外，其他接口均需要在请求头中携带 JWT Token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://production-server.com
    description: Production server
...
paths:
...
  /api/v1/knowledge/file/upload:
    post:
      tags:
        - 知识库管理
      summary: 上传文件到知识库
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 文件列表（支持多文件）
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: 请选择文件
        '500':
          description: 上传失败
  /api/v1/knowledge/contents:
    get:
      tags:
        - 知识库管理
      summary: 查询文件列表
      operationId: queryFiles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
        - name: fileName
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResult'
  /api/v1/knowledge/delete:
    delete:
      tags:
        - 知识库管理
      summary: 删除文件
      operationId: deleteFiles
      security:
        - bearerAuth: []
      parameters:
        - name: ids
          in: query
          required: true
          style: form
          explode: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /api/v1/knowledge/download:
    get:
      tags:
        - 知识库管理
      summary: 下载文件
      operationId: downloadFiles
      security:
        - bearerAuth: []
      parameters:
        - name: ids
          in: query
          required: true
          style: form
          explode: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: 下载成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'

  # ==================== 图像生成接口 ====================
...
  /api/v1/sensitive/add:
    post:
      tags:
        - 敏感词管理
      summary: 新增敏感词
      operationId: addSensitiveWord
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitiveWord'
      responses:
        '200':
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
  /api/v1/sensitive/batch:
    post:
      tags:
        - 敏感词管理
      summary: 批量删除敏感词
      operationId: deleteSensitiveWords
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
  /api/v1/sensitive/page:
    get:
      tags:
        - 敏感词管理
      summary: 分页查询敏感词
      operationId: getSensitiveWordPage
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: true
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IPage_SensitiveWord_'

  # ==================== 敏感词分类接口 ====================
  /api/v1/category/add:
    post:
      tags:
        - 敏感词分类
      summary: 新增分类
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitiveCategory'
      responses:
        '200':
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: boolean
  /api/v1/category/batch:
    delete:
      tags:
        - 敏感词分类
      summary: 批量删除分类
      operationId: batchDeleteCategories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: boolean
  /api/v1/category/update:
    put:
      tags:
        - 敏感词分类
      summary: 更新分类
      operationId: updateCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitiveCategory'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: boolean
  /api/v1/category/page:
    get:
      tags:
        - 敏感词分类
      summary: 分页查询分类
      operationId: getCategoryPage
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IPage_SensitiveCategory_'
  /api/v1/category/list:
    get:
      tags:
        - 敏感词分类
      summary: 获取全部列表
      operationId: listCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SensitiveCategory'

  # ==================== 分词统计接口 ====================
...
  /api/v1/log/page:
    get:
      tags:
        - 日志管理
      summary: 分页查询日志
      operationId: getLogInfoPage
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: methodName
          in: query
          required: false
          schema:
            type: string
        - name: className
          in: query
          required: false
          schema:
            type: string
        - name: requestParams
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IPage_LogInfo_'
  /api/v1/log/batch:
    post:
      tags:
        - 日志管理
      summary: 批量删除日志
      operationId: deleteLogInfos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          nullable: true
...
    PageResult:
      type: object
      properties:
        records:
          type: array
          items:
            type: object
        total:
          type: integer
    IPage_SensitiveWord_:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SensitiveWord'
        total:
          type: integer
        size:
          type: integer
        current:
          type: integer
        pages:
          type: integer
    IPage_SensitiveCategory_:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SensitiveCategory'
        total:
          type: integer
        size:
          type: integer
        current:
          type: integer
        pages:
          type: integer
    IPage_LogInfo_:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/LogInfo'
        total:
          type: integer
        size:
          type: integer
        current:
          type: integer
        pages:
          type: integer
    SensitiveWord:
      type: object
      properties:
        id:
          type: integer
        word:
          type: string
        categoryId:
          type: integer
        status:
          type: string
          enum: ["0", "1"]
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    SensitiveCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        createdTime:
          type: string
          format: date
        updateTime:
          type: string
          format: date
...
    LogInfo:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        className:
          type: string
        methodName:
          type: string
        requestParams:
          type: string
        responseResult:
          type: string
        executionTime:
          type: integer
        createTime:
          type: string
          format: date-time
...